Address review comments: rename to component, remove dlt_parser
- Move score/test/integration/ -> score/test/component/ to distinguish
  from nightly integration tests (reviewer: rmaddikery)
- Rename integration_config_files -> integration_test_config_files
  in score/datarouter/BUILD (reviewer: rmaddikery)
- Remove custom dlt_parser.py; refactor all tests to use DltLogRecord
  from score_itf (score.itf.plugins.dlt.dlt_window) via download_dlt()
  helper in logging_plugin.py (reviewer: rmaddikery)
