+++ score/datarouter/dlt_filetransfer_trigger_lib/BUILD
@@ -31,7 +31,7 @@ cc_library(
         "//platform/aas/pas/datarouterconf/src/services:__pkg__",
         "//platform/aas/test/pas:__subpackages__",
         "//score/datarouter/test:__subpackages__",
-        "//score/test/integration:__subpackages__",
+        "//score/test/component:__subpackages__",
         "@score_baselibs//score/logging:__pkg__",
     ],
     deps = [
