+++ score/test/component/filetransfer/etc/logging.json
