+++ score/test/component/logging_plugin.py
