+++ score/test/component/mw_log/etc/logging.json
