+++ score/test/component/mw_log/logging_app.cpp
