+++ score/test/component/mw_log_kfile/etc/logging.json
